+++ app/src/main/java/com/example/takanimali/data/CollectWasteRepositoryImpl.kt
@@ -22,7 +22,7 @@ class CollectWasteRepositoryImpl @Inject constructor(private val apiService: Api
     override suspend fun collectWaste(
         block_id: Int,
         location_id: Int,
-        quantity: Int,
+        quantity: Float,
         team_leader_id: Int,
         user_id: Int,
         waste_id: Int,
